+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - добавлен объект выписка по карточному счёту `AccountStatement`
 - добавлен объект идентификатор уровня карты `LevelId`
 - для сущности `Cards` для роли `Organization` в методе `levelUp` теперь возвращается обновлённый объект карты
+- для сущности `Cards` для роли `Organization` добавлен метод `getPaymentBalance` [issue#45](https://github.com/rarus/bonus-server-php-sdk/issues/45)
+- добавлен объект доступный для платежа баланс по карточному счёту `PaymentBalance` 
 
 ## 0.3.5 (24.07.2018)
 - для сущности `Transactions` в транспорте для роли `Organization` добавлен метод `addRefundTransaction`